added home template
