+++ src/Components/Home.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Home = () =>{
+    return (
+        <h1>Home</h1>
+    )
+}
+
+export default Home;
